+++ codegen-core/src/main/kotlin/software/amazon/smithy/rust/codegen/core/rustlang/RustWriter.kt
@@ -510,7 +510,7 @@ class RustWriter private constructor(
      * Callers must take care to use [this] when writing to ensure code is written to the right place:
      * ```kotlin
      * val writer = RustWriter.forModule("model")
-     * writer.withModule(RustModule.public("nested")) {
+     * writer.withInlineModule(RustModule.public("nested")) {
      *   Generator(...).render(this) // GOOD
      *   Generator(...).render(writer) // WRONG!
      * }
