+++ codegen-server/python/src/main/resources/META-INF/services/software.amazon.smithy.build.SmithyBuildPlugin
@@ -2,4 +2,4 @@
 # Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 # SPDX-License-Identifier: Apache-2.0
 #
-software.amazon.smithy.rust.codegen.server.python.smithy.PythonCodegenServerPlugin
+software.amazon.smithy.rust.codegen.server.python.smithy.RustServerCodegenPythonPlugin
